# Improve String.valid?/1 performance