@asakura 2024-03-12
Unfortunately I've found why exactly this algo was fast: a bunch of corner cases were missing (as well in BulkMatch32 and other variants) . The proper implementation of Branchy Range Validator algorithm implemented in Elixir is attached down below this comment. Maybe someone might find it useful.

Regarding :fast_ascii - because most of it is fast C code with ascii hack as a cherry on the top of cake I would go with that option: it have the best best performance so far and for ascii string can be just beaten by anything else except SIMD algo of course.

https://gist.github.com/asakura/71d8f754403dda46051163d6b7b5b741

```elixir
defmodule AndOr4 do
  def valid?(<<>>), do: true
  def valid?(str) when not is_binary(str), do: false
  def valid?(str), do: do_valid?(str)

  def do_valid?(<<a::56, rest::binary>>)
      when Bitwise.band(a, 0x80808080808080) == 0,
      do: do_valid?(rest)

  def do_valid?(<<a::8, rest::binary>>) do
    rest =
      cond do
        a <= 127 ->
          rest

        a >= 194 and a <= 223 ->
          case rest do
            <<a::8, rest::binary>> when a < 192 -> rest
            _ -> false
          end

        a >= 224 and a <= 239 ->
          case rest do
            <<b::16, rest::binary>>
            when Bitwise.band(b, 0x8080) == 0x8080 and
                   Bitwise.bor(b, 0xBFBF) == 0xBFBF ->
              cond do
                a == 224 ->
                  c = Bitwise.bsr(a, 8)
                  cond do
                    c >= 128 and c <= 159 -> false
                    :otherwise -> rest
                  end

                a == 237 ->
                  c = Bitwise.bsr(a, 8)
                  cond do
                    c >= 160 and c <= 191 -> false
                    :otherwise -> rest
                  end

                (a >= 225 and a <= 236) or a == 238 or a == 239 ->
                  rest

                :otherwise ->
                  false
              end

            _ ->
              false
          end

        a >= 240 ->
          case rest do
            <<b::24, rest::binary>>
            when Bitwise.band(b, 0x808080) == 0x808080 and
                   Bitwise.bor(b, 0xBFBFBF) == 0xBFBFBF ->
              cond do
                a == 240 ->
                  c = Bitwise.bsr(b, 16)
                  cond do
                    c >= 128 and c <= 143 -> false
                    :otherwise -> rest
                  end

                a >= 241 and a <= 243 ->
                  rest

                a == 244 ->
                  c = Bitwise.bsr(b, 16)
                  cond do
                    c >= 160 and c <= 191 -> false
                    :otherwise -> rest
                  end

                :otherwise ->
                  false
              end

            _ ->
              false
          end

        :otherwise ->
          false
      end

    do_valid?(rest)
  end

  def do_valid?(<<>>), do: true
  def do_valid?(false), do: false
end
```