@mtrudel 2024-03-12
That's so often the way with these perf workups; the benchmarks look great but then you realize it was really a free lunch the whole time. I've been there!

Thanks for the effort regardless @asakura !